Fix compiler warnings

+++ src/kernel/scheduler.c
@@ -53,7 +53,7 @@ void schedule(void) {
 void yield(void) { schedule(); }
 
 void scheduler_task_enumerate(task_enumerate_cb cb) {
-	for (int i = 0; i < task_count; i++) {
+	for (size_t i = 0; i < task_count; i++) {
 		cb(task_tasks[i]);
 	}
 }
